fix wrong type of data for add_options

+++ lib/bio/appl/ngs/cufflinks.rb
@@ -613,7 +613,7 @@ module Bio
         add_option "ref-gtf", :type => :string, :aliases => '-g'
         add_option "ref-sequence", :type => :string, :aliases => '-s'
         add_option "min-isoform-fraction", :type => :numeric, :aliases => '-m'
-        add_option "num-threads", :type => :integer, :aliases => '-p'
+        add_option "num-threads", :type => :numeric, :aliases => '-p'
         add_option "keep-tmp", :type => :boolean, :aliases => 't'
       end #Merge
     end #Cufflinks
